add support for k-points with lcao/blacs (which I/we apparently forgot to do)
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@7029 6419b465-d902-0410-997c-be2266baf723

+++ gpaw/blacs.py
@@ -923,7 +923,7 @@ class BlacsOrbitalLayouts(BlacsLayouts):
         nao = self.nao
         
         if rho_mM is None:
-            rho_mM = self.mMdescriptor.zeros()
+            rho_mM = self.mMdescriptor.zeros(dtype=C_nM.dtype)
         
         Cf_nM = C_nM * f_n[:, None]
         pblas_simple_gemm(self.nMdescriptor, self.nMdescriptor,
